Splitting backend.neo4j into backend.neo4j and backend.neo4j_spatial
